Move SSB back and move rands files into folders
