+++ data/mods/gen9ssb/random-teams.ts
@@ -1,4 +1,4 @@
-import RandomTeams from './gen9-teams';
+import RandomTeams from '../../random-battles/gen9/teams';
 
 export interface SSBSet {
 	species: string;
