+++ data/random-battles/gen2/sets.json
