+++ data/random-battles/gen3/sets.json
