+++ data/random-battles/gen4/sets.json
