+++ data/random-battles/gen6/factory-sets.json
