+++ data/random-battles/gen7/bss-factory-sets.json
