+++ data/random-battles/gen7letsgo/data.json
