+++ data/random-battles/gen8/bss-factory-sets.json
