+++ data/random-battles/gen8/cap-1v1-sets.json
