+++ data/random-battles/gen8/factory-sets.json
