+++ data/random-battles/gen9/sets.json
