+++ data/random-battles/sharedpower/teams.ts
@@ -1,4 +1,4 @@
-import RandomTeams from './gen9-teams';
+import RandomTeams from '../gen9/teams';
 
 export class RandomSharedPowerTeams extends RandomTeams {}
