+++ sim/teams.ts
@@ -620,6 +620,8 @@ export const Teams = new class Teams {
 		format = Dex.formats.get(format);
 		if (toID(format).includes('gen9computergeneratedteams')) {
 			TeamGenerator = require(Dex.forFormat(format).dataDir + '/cg-teams').default;
+		} else if (toID(format).includes('gen9ssb')) {
+			TeamGenerator = require(`../data/mods/gen9ssb/random-teams`).default;
 		} else {
 			TeamGenerator = require(`../data/random-battles/${format.mod}-teams`).default;
 		}
