+++ test/random-battles/gen6.js
@@ -8,11 +8,11 @@ const {testTeam, testNotBothMoves, testSet, testHiddenPower, testAlwaysHasMove, 
 
 describe('[Gen 6] Random Battle (slow)', () => {
 	const options = {format: 'gen6randombattle'};
-	const setsJSON = require(`../../dist/data/random-battles/gen6-sets.json`);
+	const setsJSON = require(`../../dist/data/random-battles/gen6/sets.json`);
 	const dex = Dex.forFormat(options.format);
 
 	describe("New set format", () => {
-		const filename = '../../data/random-battles/gen6-sets.json';
+		const filename = '../../data/random-battles/gen6/sets.json';
 		it(`${filename} should have valid set data`, () => {
 			const setsJSON = require(filename);
 			const validRoles = [
